refactor: sessionTimelineView쪽 컴포넌트 분리 ( 차트, 디테일 페이지 분리 )
